% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{distAtNodes}
\alias{distAtNodes}
\title{Distribution at Nodes}
\usage{
distAtNodes(mod, df, resp_varname)
}
\arguments{
\item{mod}{A tree model constructed by the \code{tree} package.}

\item{df}{A data frame (usually the training set, quiz set or test set).}

\item{resp_varname}{The name of the response variable, as a character string.}
}
\value{
A table object.
}
\description{
Easily find the distribtution of the response variable at each of the
nodes of a classification tree.
}
\note{
For more details on use, consult
\url{http://statistics.rainandrhino.org/tigerTree/distAtNodes.html}
}
\examples{
\dontrun{
dfs <- divideTrainTest(seed = 3030, prop.train = 0.67, data = iris)
irisTrain <- dfs$train
irisTest <- dfs$test
tr.mod <- tree(Species ~ ., data = irisTrain)
distAtNodes(tr.mod, df = irisTest, resp_varname = "Species")
}
}

