% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{divideTrainTest}
\alias{divideTrainTest}
\title{Train-Test Splitting}
\usage{
divideTrainTest(seed = NULL, prop.train = 0.7, data)
}
\arguments{
\item{seed}{A seed for randomization (recommended).}

\item{prop.train}{Proportion of the data to include in the training set.}

\item{data}{The data frame to be split.}
}
\value{
A list with elements named \code{test} and \code{train}.
}
\description{
Divide a data frame into training and test sets.
}

