% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tigerTree-package.R
\docType{package}
\name{tigerTree}
\alias{tigerTree}
\alias{tigerTree-package}
\title{Functions for Working with Trees in Elementary Data Analysis}
\description{
For use by elementary students in statistics and data analysis.
}
