% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{treeDetective}
\alias{treeDetective}
\title{Tree-Detective}
\usage{
treeDetective(mod, data, rowname = "1")
}
\arguments{
\item{mod}{A tree model constructed by the \code{tree} package.}

\item{data}{Data frame used to construct the model.}

\item{rowname}{Character indicating the initial row name in \code{mod$frame}.
Set to "1" by default.}
}
\value{
Side-effects to console.
}
\description{
Yes-No questions guide the user through a classification or regression
tree.
}
\note{
For more details on use, consult
\url{http://statistics.rainandrhino.org/tigerTree/treeDetective.html}
}
\examples{
\dontrun{
tr.mod <- tree(Species ~ ., data = iris)
treeDetective(tr.mod, iris)
}
}
