% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuneTree.R
\name{tuneTree}
\alias{tuneTree}
\title{Hand-Tune a Tree Model}
\usage{
tuneTree(formula, data, testSet, truth)
}
\arguments{
\item{formula}{formula for \code{tree}.}

\item{data}{training data to make the tree models}

\item{testSet}{quiz data to try the models on}

\item{truth}{values of response variable in the quiz data}
}
\value{
No values returned
}
\description{
Hand-Tune a Tree Model
}
\note{
For more details on use, consult
\url{http://statistics.rainandrhino.org/tigerTree/tuneTree.html}
}
