% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tryTree}
\alias{tryTree}
\title{Predcit with a Tree}
\usage{
tryTree(mod, testSet, truth, printOut = TRUE)
}
\arguments{
\item{mod}{A tree model constructed by package \code{tree}.}

\item{testSet}{The test set (a data frame).}

\item{truth}{Correct values of the response variable.}

\item{printOut}{If TRUE, provide a printout to the console.}
}
\value{
A list,  If \code{md} is a classification tree, then the list
contains elements \code{error.rate} and \code{confusion} (the confucsion matrix).
If \code{md} is a classification tree, then the list
contains elements \code{error.rate} is a regression tree, then the list contains
elements named \code{mse} and \code{rmse} (mean-square error and root mean square
error).
}
\description{
Predcit with a Tree
}
\examples{
dfs <- divideTrainTest(seed = 3030, prop.train = 0.67, data = iris)
irisTrain <- dfs$train
irisTest <- dfs$test
tr.mod <- tree(Species ~ ., data = iris)
summary(tr.mod)
tryTree(mod = tr.mod, testSet = irisTest, truth = irisTest$Species)
}

